+++ src/main/kotlin/com/github/bartjaskulski/wordpresstextdomain/util/FunctionReferenceBag.kt
@@ -1,0 +1,12 @@
+package com.github.bartjaskulski.wordpresstextdomain.util
+
+import com.jetbrains.php.lang.psi.elements.FunctionReference
+import com.jetbrains.php.lang.psi.elements.ParameterList
+
+class FunctionReferenceBag(
+    val parameterList: ParameterList,
+    val functionReference: FunctionReference,
+    val parameterBag: ParameterBag
+) {
+
+}