+++ src/main/kotlin/com/github/bartjaskulski/wordpresstextdomain/util/ParameterBag.kt
@@ -1,0 +1,11 @@
+package com.github.bartjaskulski.wordpresstextdomain.util
+
+import com.intellij.psi.PsiElement
+
+class ParameterBag(
+    val index: Int,
+    val psiElement: PsiElement
+) {
+
+    val value = PhpElementsUtil().getFunctionParameter(psiElement)
+}