+++ src/test/kotlin/com/github/bartjaskulski/wordpresstextdomain/MyPluginTest.kt
@@ -1,39 +1,20 @@
 package com.github.bartjaskulski.wordpresstextdomain
 
-import com.intellij.ide.highlighter.XmlFileType
-import com.intellij.openapi.components.service
-import com.intellij.psi.xml.XmlFile
 import com.intellij.testFramework.TestDataPath
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
-import com.intellij.util.PsiErrorElementUtil
-import com.github.bartjaskulski.wordpresstextdomain.services.MyProjectService
 
 @TestDataPath("\$CONTENT_ROOT/src/test/testData")
 class MyPluginTest : BasePlatformTestCase() {
 
-    fun testXMLFile() {
-        val psiFile = myFixture.configureByText(XmlFileType.INSTANCE, "<foo>bar</foo>")
-        val xmlFile = assertInstanceOf(psiFile, XmlFile::class.java)
-
-        assertFalse(PsiErrorElementUtil.hasErrors(project, xmlFile.virtualFile))
-
-        assertNotNull(xmlFile.rootTag)
-
-        xmlFile.rootTag?.let {
-            assertEquals("foo", it.name)
-            assertEquals("bar", it.value.text)
-        }
+    fun testCompletion() {
+//        myFixture.configureByFiles("domain/domain.php", "domain/translatable.php");
+//        myFixture.copyDirectoryToProject("domain", "domain")
+//        myFixture.complete(CompletionType.BASIC);
+        val lookupElementStrings = myFixture.lookupElementStrings;
+        assertNotNull(lookupElementStrings);
+        lookupElementStrings?.contains("text-domain")?.let { assertTrue(it) };
+//        assertSameElements(lookupElementStrings, "text-domain", "language", "message", "tab", "website");
     }
 
-    fun testRename() {
-        myFixture.testRename("foo.xml", "foo_after.xml", "a2")
-    }
-
-    fun testProjectService() {
-        val projectService = project.service<MyProjectService>()
-
-        assertEquals(4, projectService.getRandomNumber())
-    }
-
-    override fun getTestDataPath() = "src/test/testData/rename"
+    override fun getTestDataPath() = "src/test/testData"
 }