+++ src/test/testData/domain/domain.php
@@ -1,0 +1,4 @@
+<?php
+/**
+ * Text Domain: text-domain
+ */