+++ src/test/testData/domain/translatable.php
@@ -1,0 +1,3 @@
+<?php
+
+__('Translatable', '<caret>');